Fixed hover slide at redeindividual and added responsive size of sidebar
The original hover effect had a bug when you scrolled the sidebar. It wouldn't stay in position with the original element and would go AWHIRE

Added 3 types of  responsivity for smaller, older screens that increases the sidebar width responsively along with the viewport.
For this I created media queries for the viewport.
Also had to change the hover div width so that it would fit the sidebar independently of the sidebar width