Changes to make single instance possible
This changes are so that webwindow will not start if there is another webwindow open inside the same process as this class is static and doesn't instantiate into objects so it needed a boolean to manage if it was the only instance open.

If there is another instance open it just fetches that instance window handle and brings it up in windows.