Services changes to get up to date with blazor
In a new blazor wasm update the way that services were added to the blazor builder changed and the startup progress for webwindow also had to change. 

I basically made it so the startup method for webwindow recieves a service collection making it simpler to add existing and new services to later inject on the blazor components.
this doesn't remove the need for the staartup task needed to start up webwindow