Fixed the reopening bug and more null changes
Fixed a bug in which when closing and reopening the webwindow inside the same process the navigation manager stopped working. This was due to between closing the webwindow and reopening the DesktopNavigationManager didn't close and was kept the same due to it being static so it oppened on the page that was previously open but the link where it thought is was was the beginning link.

To fix this i had to unprotect the DesktopNavigationManagerClass and then every time the WebWindow Closed i needed to instantiate a new instance of the DesktopNavigationManager in its instance object so there will not be any further coonfusion between  closing and reopening the webwindow