Custom app context
This class consists in the custom app context which will bind the process lifetime to the system tray icon.... 

It basically instantiates a new notifyicon object which is the system tray icon. Then it will give a context menu which has the Exit and open options. The exit quits the program and the open opens the webwindow. Then it gets set to be visible and, since it is beeing instatiated when opening the app, it also opens the webwindow