Pipes stuff and System Tray Icon start up
Basically these files set the pipe server, the pipe client, and the pipe payload to be used later.

I used this method: https://www.autoitconsulting.com/site/development/single-instance-winform-app-csharp-mutex-named-pipes/

I needed to use pipes because, for integrity of the files, there can only be one instance of the program at a time. So basically it checks if there is an instance already open on the computer through mutex, then if there wasn't it starts a pipe server and starts up the system tray icon, this pipe server will be listening for messages on that pipe. when it detects a message it tries to open a new webwindow (Then the webwindow will decide if there is a webwindow already open and if so it will bring it up, if it isn't open, a new webwindow will be oppened. This logic is handled on the webwindow.blazor ComponentsDesktop.cs file).

If the mutex detects that there already is an instance of the app opened then it will start up a pipe client which will send a message to the already opened pipe server so that it can deal with that message. after sending that message the application closes as there is no more need for it to be opened.


When starting up the System Tray Icon it will do the default process for starting up a windows forms but instead of using a normal form to run it will use the custom application context made previously